feat: component namespace
One can overwrite the namespace name that is generated for the component:

- function `getComponentNamespace(vm)`

Should return: `{ namespace: ‘new-namespace’, loadNamespace: true}`
`namespace`: is the new namespace to use
`loadNamespace`: tells if the namespace should be loaded. Be aware that if the global options `loadComponentNamespace === false` then the namespace will not be loaded even if `loadNamespace === true`.